+++ test/Utility/Test_ConvexHull3D.cpp
@@ -6,53 +6,7 @@
 #include "Utility/ConvexHull3D.h"
 #include <string>
 #include <sstream>
-
-bool compareListVector3(const std::vector<Eigen::Vector3d> &A, const std::vector<Eigen::Vector3d> &B)
-{
-
-    if(A.size() != B.size()) return false;
-
-    for(int id = 0; id < A.size(); id++)
-    {
-        int jd = 0;
-        for(; jd < B.size(); jd++){
-            if(Approx(0) == (A[id] - B[jd]).norm()){
-                break;
-            }
-        }
-
-        if(jd == B.size())
-            return false;
-    }
-    return true;
-
-}
-
-void string2ListVector3(std::string data, std::vector<Eigen::Vector3d> &pts)
-{
-
-    std::stringstream iss(data);
-    double number;
-    int index = 0;
-    while(iss >> number){
-
-        if(index % 3 == 0){
-            pts.push_back(Eigen::Vector3d(0, 0, 0));
-        }
-
-        pts.back()[index % 3] = number;
-        index += 1;
-    }
-
-    return;
-}
-
-void printListVector3(std::vector<Eigen::Vector3d> &list){
-    for(Eigen::Vector3d pt : list){
-        std::cout << pt.transpose() << std::endl;
-    }
-}
-
+#include "ListVector3.h"
 TEST_CASE("ConvexHull3D")
 {
     std::string data = "0.3215426810286406 0.1678336189760208 -0.2203710966001927 \n"
@@ -118,14 +72,14 @@ TEST_CASE("ConvexHull3D")
                          "0.2100642609503719 -0.4499717643018549 0.3245569875692548";
 
     ConvexHull3D<double>::ListVector3 pts,resPts;
-    string2ListVector3(data, pts);
-    string2ListVector3(result, resPts);
+    convertString3ToListVector3(data, pts);
+    convertString3ToListVector3(result, resPts);
 
     ConvexHull3D<double> convexhull3d;
 
     ConvexHull3D<double>::ListVector3 outPts;
     ConvexHull3D<double>::ListVector3i outTri;
-    convexhull3d.computeQuickHull(pts, outPts, outTri);
+    convexhull3d.compute(pts, outPts, outTri);
 
     REQUIRE(compareListVector3(outPts, resPts) == true);